feat(components/plexloginbutton): added PlexLoginButton
